feat: long task test
