+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "tsTojs": "tsc --build tsconfig.json"
   },
   "repository": {
     "type": "git",
@@ -25,6 +26,10 @@
   },
   "homepage": "https://github.com/wuyunqiang/blog#readme",
   "devDependencies": {
-    "eslint": "^8.41.0"
+    "@types/node": "^20.3.0",
+    "eslint": "^8.41.0",
+    "ts-node": "^10.9.1",
+    "tslib": "^2.5.3",
+    "typescript": "^5.1.3"
   }
 }
