+++ src/task.js
@@ -157,82 +157,5 @@ export const nextFrameExecute = async (task) => {
   return p;
 };
 
-/**
- * 支持动态添加
- * 支持并行&串行
- * 时间切片
- */
 
-export const TaskStatus = {
-  Done: 1,
-  Processing: 2,
-};
-export class DynamicTasks {
-  status = TaskStatus.Done;
-  store = {}; // 存储结果
-  handles = []; // 回调列表
-  parallelMax;
-  frame = false;
 
-  constructor(config = {}) {
-    this.parallelMax = config.parallelMax || 3;
-    this.frame = config.frame;
-  }
-  getResult() {
-    return this.store;
-  }
-  removeItem(key) {
-    if (this.store[key]) {
-      Reflect.deleteProperty(this.store, key);
-    }
-  }
-  removeStore() {
-    this.store = {};
-  }
-  add(data) {
-    const list = Array.isArray(data) ? data : [data];
-    this.handles.push(...list);
-  }
-
-  async loopRun(_resolve) {
-    if (!this.handles.length) {
-      _resolve(this.store);
-      return;
-    }
-    const { task, key } = this.handles.shift();
-    try {
-      const runner = this.frame
-        ? () => nextFrameExecute(() => task(this.store))
-        : () => task(this.store);
-      const res = await runner();
-      this.store[key] = {
-        status: "succ",
-        data: res,
-      };
-    } catch (error) {
-      this.store[key] = {
-        status: "fail",
-        data: error,
-      };
-    } finally {
-      this.loopRun(_resolve);
-    }
-  }
-  async run() {
-    let _resolve;
-    const p = new Promise((resolve) => {
-      _resolve = resolve;
-    });
-    if (!this.handles.length) {
-      this.status = TaskStatus.Done;
-      _resolve(this.store);
-      return p;
-    }
-    if (this.status == TaskStatus.Processing) {
-      return p;
-    }
-    this.status = TaskStatus.Processing;
-    this.loopRun(_resolve);
-    return p;
-  }
-}
